--- README.md
@@ -1,12 +1,0 @@
-# Syncher
-
-Module-wrapper under standart "rsync" linux function.
-
-syncher.interface(cli="", files="", user="", port="", host="", dist=""),
-
-* cli - 
-* files - 
-* user - 
-* port - 
-* host - 
-* dist -
