--- interface_test.py
@@ -1,3 +1,0 @@
-import syncher
-
-syncher.interface(cli="-avz", files="./123 ./1* ", user="root", host="192.168.56.102", dist="/root")
