+++ try_all_by_console.py
